+++ app/src/main/java/com/amsterdam/cutetudee/data/service/TaskServiceImpl.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.data.repository
+package com.amsterdam.cutetudee.data.service
 
 import com.amsterdam.cutetudee.data.local.dao.TaskDao
 import com.amsterdam.cutetudee.data.mapper.toLong
@@ -6,7 +6,7 @@ import com.amsterdam.cutetudee.data.mapper.toTask
 import com.amsterdam.cutetudee.data.mapper.toTaskEntity
 import com.amsterdam.cutetudee.data.mapper.toTaskListFlow
 import com.amsterdam.cutetudee.domain.model.Task
-import com.amsterdam.cutetudee.domain.repository.TaskService
+import com.amsterdam.cutetudee.domain.service.TaskService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.datetime.LocalDate
 import kotlin.uuid.ExperimentalUuidApi
