+++ app/src/main/java/com/amsterdam/cutetudee/di/DataModule.kt
@@ -1,10 +1,10 @@
 package com.amsterdam.cutetudee.di
 
 import com.amsterdam.cutetudee.data.local.roomDB.TudeeDatabase
-import com.amsterdam.cutetudee.data.repository.CategoryServiceImpl
-import com.amsterdam.cutetudee.data.repository.TaskServiceImpl
-import com.amsterdam.cutetudee.domain.repository.CategoryService
-import com.amsterdam.cutetudee.domain.repository.TaskService
+import com.amsterdam.cutetudee.data.service.CategoryServiceImpl
+import com.amsterdam.cutetudee.data.service.TaskServiceImpl
+import com.amsterdam.cutetudee.domain.service.CategoryService
+import com.amsterdam.cutetudee.domain.service.TaskService
 import org.koin.android.ext.koin.androidApplication
 import org.koin.dsl.module
