+++ app/src/main/java/com/amsterdam/cutetudee/domain/service/CategoryService.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.domain.repository
+package com.amsterdam.cutetudee.domain.service
 
 import com.amsterdam.cutetudee.domain.model.Category
 import kotlinx.coroutines.flow.Flow
